refactor: update djinn-server structure
