--- djinn-rs/djinn-server/src/handlers/mod.rs
@@ -1,3 +1,0 @@
-mod complete;
-
-pub use complete::*;
